# swagger.yaml
openapi: 3.0.0
info:
  title: E-Commerce API
  contact: {}
  version: '1.0'
servers:
- url: http://3.220.174.105:8080/api/v1
  variables: {}
- url: https://kd2qzwtqf2.execute-api.us-east-1.amazonaws.com/Test/deleteimage
  variables: {}
paths:
...
  /products:
    post:
      tags:
      - Products
      summary: Create Product
      operationId: CreateProduct
      parameters: []
      requestBody:
        content:
          multipart/form-data:
            encoding: {}
            schema:
              required:
              - name
              - price
              - description
              - category
              - company
              - inventory
              type: object
              properties:
                name:
                  type: string
                price:
                  type: string
                description:
                  type: string
                category:
                  type: string
                company:
                  type: string
                inventory:
                  type: string
        required: false
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
      servers:
      - url: http://3.220.174.105:8080/api/v1
        variables: {}
    get:
      tags:
      - Products
      summary: Get All Products
      operationId: GetAllProducts
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
      servers:
      - url: http://3.220.174.105:8080/api/v1
        variables: {}
  /products/{id}:
    patch:
      tags:
      - Products
      summary: Update Product
      operationId: UpdateProduct
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
      servers:
      - url: http://3.220.174.105:8080/api/v1
        variables: {}
    delete:
      tags:
      - Products
      summary: Delete Product
      operationId: DeleteProduct
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
      servers:
      - url: http://3.220.174.105:8080/api/v1
        variables: {}
  /products/{id}/uploadImage:
    post:
      tags:
      - Products
      summary: Upload Image
      operationId: UploadImage
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      requestBody:
        content:
          multipart/form-data:
            encoding: {}
            schema:
              required:
              - image
              type: object
              properties:
                image:
                  type: string
                  format: binary
        required: false
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
      servers:
      - url: http://3.220.174.105:8080/api/v1
        variables: {}
...